Remove console logs, update component references, and refactor drag and drop logic
