+++ apps/application/packages/components/Text/default.json
@@ -2,7 +2,7 @@
   "configuration": {
     "padding": 0,
     "fontSize": 25,
-    "color": "orange",
+    "color": "yellow",
     "paddingX": 10,
     "paddingY": 10,
     "backgroundColor": "green",
